+++ assets/r/explore_visual_numeric.R
@@ -1,11 +1,11 @@
-# Rattle Scripts: Visual Displays
+# Rattle Scripts: For dataset `ds` generate useful plots of numeric variable.
 #
 # Copyright (C) 2024, Togaware Pty Ltd.
 #
 # License: GNU General Public License, Version 3 (the "License")
 # https://www.gnu.org/licenses/gpl-3.0.en.html
 #
-# Time-stamp: <Friday 2024-11-29 07:28:12 +1100 Graham Williams>
+# Time-stamp: <Tuesday 2024-12-10 17:20:42 +1100 Graham Williams>
 #
 # Licensed under the GNU General Public License, Version 3 (the "License");
 #
@@ -22,10 +22,8 @@
 # You should have received a copy of the GNU General Public License along with
 # this program.  If not, see <https://www.gnu.org/licenses/>.
 #
-# Author: Graham Williams
+# Author: Graham Williams, Kevin Wang
 
-# Visual presentation of variables.
-#
 # TIMESTAMP
 #
 # References:
@@ -34,27 +32,39 @@
 #
 # https://survivor.togaware.com/datascience/ for further details.
 
+# Load required packages from the local library into the R session.
+
 library(dplyr)
 library(ggplot2)
 library(rattle)
 
+# Preprocess the dataset to generate a temporary dataset `tds` for the
+# plots.
+
+# If we want to ignore missing values for the group-by then we filter
+# out rows in `ds` with the group-by variable having `NA`. Otherwise
+# simply use the whole dataset `ds`.
+
+if (IGNORE_MISSING_GROUP_BY) {
+  tds <- dplyr::filter(ds, !is.na(GROUP_BY_VAR))
+} else {
+  tds <- ds
+}
+
+# Ensure the group-by variable is a factor in out temporary dataset.
+
+tds <- dplyr::mutate(tds, GROUP_BY_VAR=as.factor(GROUP_BY_VAR))
+
 ########################################################################
 # BOX PLOT
 ########################################################################
-
-# TODO 20241120 gjw ADD ALL TO LEGEND - ACTUALLY REMOVE LEGEND AS ADDS NO VALUE
-
-# Preprocess the dataset for the plot.
-
-tds <- dplyr::mutate(ds, GROUP_BY_VAR=as.factor(GROUP_BY_VAR))
-
-# When the confidence interval for the notch of the boxplot extends
-# beyond the hinges (upper or lower limits of the boxplot) the notches
-# go outside the hinges. To avoid this precompute the values for the
-# notches and compare them to the hinges. The following does it for
-# one variable. Need for all. For now 20241121 add an option to turn
-# the notch on and off.
-
+## # 20241210 gjw When the confidence interval for the notch of the
+## # boxplot extends beyond the hinges (upper or lower limits of the
+## # boxplot) the notches go outside the hinges. To avoid this we can
+## # precompute the values for the notches and compare them to the
+## # hinges. The following does it for one variable. For now we will set
+## # notch to TUE or FALSE but eventually we want to use is_notch_safe.
+##
 ## is_notch_safe <- function(data, x, y) {
 ##   summary_data <- data %>%
 ##     group_by({{ x }}) %>%
@@ -66,16 +76,14 @@ tds <- dplyr::mutate(ds, GROUP_BY_VAR=as.factor(GROUP_BY_VAR))
 ##     )
 ##   any(summary_data$notch_lower < summary_data$ymin | summary_data$notch_upper > summary_data$ymax)
 ## }
-
+##
 ## # Check if the notch is safe.
-
+##
 ## notch_safe <- is_notch_safe(tds, , hwy)
-
 
 # Display box plot for the selected variable.
 
 svg("TEMPDIR/explore_visual_boxplot.svg", width=10)
-
 tds %>%
   ggplot2::ggplot(ggplot2::aes(y=SELECTED_VAR)) +
   ggplot2::geom_boxplot(ggplot2::aes(x="All"), notch=BOXPLOT_NOTCH, fill="grey") +
@@ -85,18 +93,15 @@ tds %>%
   ggplot2::xlab(paste("GROUP_BY_VAR\n\n", paste("TIMESTAMP", username), sep="")) +
   ggplot2::ggtitle("Distribution of SELECTED_VAR by GROUP_BY_VAR") +
   SETTINGS_GRAPHIC_THEME()
-
 dev.off()
 
 ########################################################################
 # DENSITY
 ########################################################################
-
 # TODO 20241120 gjw ADD ALL TO LEGEND
 
 svg("TEMPDIR/explore_visual_density.svg", width=10)
-
-ds %>%
+tds %>%
   dplyr::mutate(GROUP_BY_VAR=as.factor(GROUP_BY_VAR)) %>%
   dplyr::select(SELECTED_VAR, GROUP_BY_VAR) %>%
   ggplot2::ggplot(ggplot2::aes(x=SELECTED_VAR)) +
@@ -106,7 +111,6 @@ ds %>%
   ggplot2::ggtitle("Distribution of SELECTED_VAR by GROUP_BY_VAR") +
   ggplot2::labs(fill="GROUP_BY_VAR", y="Density") +
   SETTINGS_GRAPHIC_THEME()
-
 dev.off()
 
 ########################################################################
@@ -114,8 +118,7 @@ dev.off()
 ########################################################################
 
 svg("TEMPDIR/explore_visual_ecdf.svg", width=10)
-
-ds %>%
+tds %>%
   dplyr::mutate(GROUP_BY_VAR=as.factor(GROUP_BY_VAR)) %>%
   dplyr::select(SELECTED_VAR, GROUP_BY_VAR) %>%
   ggplot2::ggplot() +
@@ -127,11 +130,10 @@ ds %>%
   ggplot2::ggtitle("Empirical Cumulative Distribution of SELECTED_VAR by GROUP_BY_VAR") +
   ggplot2::labs(fill="GROUP_BY_VAR", y=expression("ECDF - Proportion <= x")) +
   SETTINGS_GRAPHIC_THEME()
-
 dev.off()
 
 ########################################################################
-# BENFORD'S LAW 
+# BENFORD'S LAW
 ########################################################################
 
 # Initialies the parameters.
@@ -166,10 +168,8 @@ len <- nchar(as.character(tds[1, 1]))
 
 # Plot the digital distribution
 
-svg("TEMPDIR/explore_visual_benford.svg", width=10)
-
 p <- ggplot2::ggplot(dsm,
-                     ggplot2::aes_string(x      = "digit", 
+                     ggplot2::aes_string(x      = "digit",
                                          y      = "value",
                                          colour = "variable",
                                          shape  = "variable")) +
@@ -183,6 +183,7 @@ if (substr(as.character(tds[1, 1]), 1, 1) == "1") {
   10^(len - 1)))
   } else {p <- p + ggplot2::scale_x_continuous(breaks = seq(0, 9, 1))}
 
+svg("TEMPDIR/explore_visual_benford.svg", width=10)
 p +
   ggplot2::ylab("Frequency (%)") +
   ggplot2::xlab("Digits") +
@@ -190,19 +191,18 @@ p +
   ggtitle("Digital Analysis of First Digit of SELECTED_VAR by GROUP_BY_VAR") +
   ggplot2::xlab(paste("Digits\n\n", paste("TIMESTAMP", username), sep="")) +
   SETTINGS_GRAPHIC_THEME()
-
 dev.off()
 
-########################################################################
-# PAIRS - REQUIRES TWO VARIABLES
-########################################################################
-
-# Display a pairs plot for the selected variables. 
-
-# Use GGally's ggpairs() to do the hard work.
-
+## ########################################################################
+## # PAIRS - REQUIRES TWO VARIABLES
+## ########################################################################
+##
+## # Display a pairs plot for the selected variables.
+##
+## # Use GGally's ggpairs() to do the hard work.
+##
 ## svg("TEMPDIR/explore_visual_pairs.svg", width=10)
-
+##
 ## ds %>%
 ##   dplyr::mutate(GROUP_BY_VAR=as.factor(GROUP_BY_VAR)) %>%
 ##   GGally::ggpairs(columns=c(3,4),
@@ -218,10 +218,8 @@ dev.off()
 ##                 legend=3) +
 ##   ggplot2::theme(panel.grid.major=ggplot2::element_blank(), legend.position="right") +
 ##   ggplot2::xlab(paste("\n\n", "TIMESTAMP", sep=""))
-
-  
-# ggplot2::scale_alpha_continuous(guide=FALSE) +
-#  ggplot2::scale_fill_brewer(palette=rattlePalette) +
-#  ggplot2::scale_colour_brewer(palette=rattlePalette)
-
-# dev.off()
+##
+## # ggplot2::scale_alpha_continuous(guide=FALSE) +
+## #  ggplot2::scale_fill_brewer(palette=rattlePalette) +
+## #  ggplot2::scale_colour_brewer(palette=rattlePalette)
+## dev.off()